w3c-reports/12092014/mac-wk2/runned-tests/dom/ranges/Range-extractContents-expected.txt

CONSOLE MESSAGE: line 49: TypeError: undefined is not a function (evaluating 'iframe.contentWindow.setupRangeTests()')
FAIL: Timed out waiting for notifyDone to be called
